weather_analysis/requirements.txt
Django
mysqlclient
python-dotenv
djangorestframework
django-cors-headers
requests
psycopg2-binary
dj-database-url
whitenoise[brotli]
gunicorn
uvicorn